Rename module and package name
Remove hyphenation
