+++ README.md
@@ -42,8 +42,8 @@ Follow these steps:
 
 2. Open an terminal and navigate to `./build/package/<platform of your choice>`
 
-3. Activate the app `go-react` (for macOS and Linux) or `go-react.exe` (for windows)
-NOTE: By default the app will run require port 80 to be available. If you need the app to use other ports type the command `go-react --help`.
+3. Activate the app `goreact` (for macOS and Linux) or `goreact.exe` (for windows)
+NOTE: By default the app will run require port 80 to be available. If you need the app to use other ports type the command `goreact --help`.
 
 4. Assuming you have no problem with STEP 3, open a browser with url to `localhost`.
