+++ go.mod
@@ -1,10 +1,10 @@
-module go-web
+module goweb
 
 go 1.13
 
 require (
 	github.com/GeertJohan/go.rice v1.0.0
 	github.com/gorilla/mux v1.7.3
-	github.com/paulwizviz/go-react v0.0.0-20200214155111-6859bedd86dd
+	github.com/paulwizviz/go-react v0.0.0-20200214155111-6859bedd86dd // indirect
 	github.com/spf13/cobra v0.0.5
 )
